No need to convert to a byte value.

+++ PacketWrapper/src/main/java/com/comphenix/packetwrapper/Packet12Animation.java
@@ -104,7 +104,7 @@ public class Packet12Animation extends AbstractPacket {
      * @return The current Animation
     */
     public int getAnimation() {
-        return handle.getIntegers().read(1).byteValue();
+        return handle.getIntegers().read(1);
     }
     
     /**
